芜湖 项目部分完工~

+++ src/analysis/src/main/java/com/tgu/team04/analysis/entity/Book_DataReview.java
@@ -7,7 +7,7 @@ public class Book_DataReview {
 
     private String press;
 
-    private Integer reviews;
-    private Integer shortComms;
+    private Float reviews;
+    private Float shortComms;
 
 }
